🎯 feat: group error handling and add CaptchaError with test
